Changed type of a couple local variables in LogCodec::cstyle() to size_t, to eliminate a compiler warning. Modified checkEventAssignmentValues to get rid of a compiler warning, and to use a more-than-32-bit integer to test Event::getUBigInt().

+++ platform/codecs/LogCodec.hpp
@@ -461,7 +461,7 @@ inline char * LogCodec::cstyle(char *cstring)
 {
 	char *ptr = cstring;
 	size_t len = strlen(cstring);
-	int num, nlen;
+	size_t num, nlen;
 
 	while ( (ptr = strchr(ptr, '\\')) ) {
 		nlen = 1;
